Object_Detection_Deployment/requirements.txt
numpy=1.17.3
pytorch-cpu=1.0.1
matplotlib=3.1.2
matplotlib-base=3.1.2
flask=1.1.1
requests=2.19.1
tqdm=4.38
pillow
torchvision-cpu